[refactor] move fastx core types and authority into separate crates
Separate out the fastx base types (used by clients, the adapter, authority, ...) and the authority code. A future PR connecting the `fastx-programmability` and `fastpay-core` crates leads to cyclic dependencies without this refactor.
