+++ Cargo.toml
@@ -3,7 +3,8 @@ members = [
         "fastpay_core",
         "fastpay",
         "fastx_programmability/adapter",
-        "fastx_programmability/framework"
+        "fastx_programmability/framework",
+	"fastx_types",
 ]
 
 [profile.release]
