+++ fastx_programmability/adapter/Cargo.toml
@@ -29,6 +29,7 @@ move-vm-runtime = { git = "https://github.com/diem/diem", rev="661a2d1367a64a020
 fastpay_core = { path = "../../fastpay_core" }
 fastx-framework = { path = "../framework" }
 fastx-verifier = { path = "../verifier" }
+fastx-types = { path = "../../fastx_types" }
 
 [dev-dependencies]
 datatest-stable = "0.1.1"
