+++ fastx_programmability/verifier/Cargo.toml
@@ -14,3 +14,5 @@ publish = false
 move-binary-format = { git = "https://github.com/diem/diem", rev="661a2d1367a64a02027e4ed8f4b18f0a37cfaa17" }
 
 fastpay_core = { path = "../../fastpay_core" }
+fastx-types = { path = "../../fastx_types" }
+
