+++ fastx_programmability/verifier/src/dummy_verifier.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Mysten Labs
 // SPDX-License-Identifier: Apache-2.0
 
-use fastpay_core::error::FastPayResult;
+use fastx_types::error::FastPayResult;
 use move_binary_format::file_format::CompiledModule;
 
 pub fn verify_module(_: &CompiledModule) -> FastPayResult {
