+++ fastx_types/Cargo.toml
@@ -1,18 +1,23 @@
 [package]
-name = "fastpay_core"
+name = "fastx-types"
 version = "0.1.0"
-authors = ["Libra <oncall+libra@xmail.facebook.com>"]
+authors = ["Mysten Labs <build@mystenlabs.com>"]
 publish = false
 edition = "2018"
 
 [dependencies]
+base64 = "0.12.3"
+bcs = "0.1.3"
+bincode = "1.3.1"
 failure = "0.1.8"
 futures = "0.3.5"
 rand = "0.7.3"
 serde = { version = "1.0.115", features = ["derive"] }
 tokio = { version = "0.2.22", features = ["full"] }
+ed25519 = { version = "1.0.1"}
+ed25519-dalek = { version = "1.0.1", features = ["batch", "serde"] }
+serde-name = "0.1.2"
+structopt = "0.3.21"
 
-fastx-types = { path = "../fastx_types" }
 move-binary-format = { git = "https://github.com/diem/diem", rev="661a2d1367a64a02027e4ed8f4b18f0a37cfaa17" }
 move-core-types = { git = "https://github.com/diem/diem", rev="661a2d1367a64a02027e4ed8f4b18f0a37cfaa17" }
-
