+++ fastx_types/src/error.rs
